active

+++ tpl/tpl-index.php
@@ -21,17 +21,27 @@
           <input type="search" placeholder="Search"/>
         </div>
       </div> -->
+      <style>
+        .main .nav .menu ul li.active a{
+         text-decoration: none;
+          color: #79BC46;
+           }
+      </style>
       <div class="menu">
         <div class="title">folders</div>
         <ul class="folders-list">
-            <?php
-            foreach($folders as $folder):?>
-          <li> 
+          
+          <li class="active">
+          <li class="<?= isset($_GET['folder_id']) ? '' : 'active' ; ?>"><a style="text-decoration: none;" href="?folder_id ?>"> <i class="fa fa-folder"></i> All Folder</a></li> 
+          </li>
+     
+            <?php foreach($folders as $folder):?>
+          <li class="active"> 
+          <li class="<?= ($_GET['folder_id'] == $folder->id) ? 'active' : '' ; ?>">
           <a style="text-decoration: none;" href="?folder_id=<?= $folder->id ?>"><i class="fa fa-folder"></i><?= $folder->name ?></a>
-          <a style="text-decoration: none;float:right;color:red;padding: right 20px;" href="?delete_folder=<?= $folder->id ?>">!!!</a>
+          <a style="text-decoration: none;float:right;color:red;padding: right 20px;" href="?delete_folder=<?= $folder->id ?>" onclick="return confirm('Are You sure delete for folder')" >!!!</a>
           </li>
           <?php endforeach; ?>
-          <li class="active"> <i class="fa fa-folder"></i>curent folder</li>
           <!-- <li><i class="fa fa-signal"></i>Activity</li>
           <li class="active"> <i class="fa fa-tasks"></i>Manage Tasks</li>
           <li> <i class="fa fa-envelope"></i>Messages</li> --> 
@@ -53,27 +63,36 @@
     <div>
 
     <div class="view">
-      <div class="viewHeader">
-        <div class="title">Manage Tasks</div>
-        <div class="functions">
-          <div class="button active">Add New Task</div>
-          <div class="button">Completed</div>
-          <div class="button inverz"><i class="fa fa-trash-o"></i></div>
-        </div>
-      </div>
-
+        
+                <div class="viewHeader">
+                  <input  style="height: 40;width: 270;"  type="text" id="addTaskInput" placeholder="Name Task" >
+                      <div class="functions">
+                           <button class="button active" id="addTaskBtn" ><div >Add New Task</div></button> 
+                        <div class="button">Completed</div>
+                      <a href="?task_id=<?= $task->id ?>"> <div class="button inverz"><i class="fa fa-trash-o"></i></div> </a>  
+                      </div>
+                </div>
+         
       <div class="content">
         <div class="list">
           <div class="title">Today</div>
           <ul>
+            <?php if(sizeof($tasks)):  ?>
             <?php foreach($tasks as $task): ?>
-            <li class="checked"><i class="fa fa-check-square-o"></i><span><?= $task->title?></span>
+            <li class="<?= $task->is_done ? 'checked' : '' ; ?>">
+            <i class="fa <?= $task->is_done ? 'fa-check-square-o' : 'fa-square-o' ; ?> "></i>
+            <span><?= $task->title ?></span>
               <div class="info">
-                <div class="button green">In progress</div><span><?= $task->created_at ?></span>
+                <span><?= $task->created_at?></span>
+                <a class="remove" style="text-decoration: none;float:right;color:red;padding: right 20px;" href="?delete_task=<?= $task->id ?>" onclick="return confirm('Are You sure delete for task')" >!!!</a>
               </div>
             </li>
+
             <?php endforeach; ?>
-            <li><i class="fa fa-square-o"></i><span>Design a new logo</span>
+              <?php else: ?>
+                <li>No Task Here ...</li>
+                <?php endif; ?>
+            <!-- <li><i class="fa fa-square-o"></i><span>Design a new logo</span>
               <div class="info">
                 <div class="button">Pending</div><span>Complete by 10/04/2014</span>
               </div>
@@ -88,7 +107,8 @@
           <ul>
             <li><i class="fa fa-square-o"></i><span>Find front end developer</span>
               <div class="info"></div>
-            </li>
+            </li> -->
+
           </ul>
         </div>
       </div>
@@ -98,30 +118,36 @@
 <!-- partial -->
   <script src='//cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js'></script><script  src="/assets/js/script.js"></script>
    <script>
-    $(document).ready(function(e){
+  
+ $(document).ready(function(e){
 
-        $("#addFolderBtn").click(function(e){
-            var input = $('input#addFolderInput');
-            $.ajax({
-                url : "proccess/ajaxHandler.php",
-                method : "post",
-                data : {action: "addFolder",folderName: input.val()},
-                success : function(response){
+  $("#addFolderBtn").click(function(e){
+      var input = $('input#addFolderInput');
+      $.ajax({
+          url : "proccess/ajaxHandler.php",
+          method : "post",
+          data : {action: "addFolder",folderName: input.val()},
+          success : function(response){
 
-                   if(response == '1'){
-              
-                    $(' <li><a style="text-decoration: none;" href="><i class="fa fa-folder"></i>'+input.val()+'</a></li>').appendTo('ul.folders-list');
-                 
-                  }else{
+            if(response == '1'){
+        
+              $(' <li><a style="text-decoration: none;" href="><i class="fa fa-folder"></i>'+input.val()+'</a></li>').appendTo('ul.folders-list');
+          
+            }else{
 
-                    alert(response);
+              alert(response);
 
-                   }
-                    
-                }
-            });
-        });
+            }  
+        }
     });
+  });
+});
+
+
+
+
+
+
    </script>
 </body>
 </html>
